Lovelace Custom Cards: Big Number Card
Lovelace is a new approach to define your user interface for
Home Assistant. We offer a lot of built-in cards, but you're not
just limited to the ones that we decided to include in the
Lovelace UI. You can build and use your own!

https://developers.home-assistant.io/docs/en/lovelace_custom_card.html

bignumber-card.js taken from Marius -> ciotlosm @ github:
https://github.com/ciotlosm/custom-lovelace/tree/master/bignumber-card

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>